Created board view
Board game FXML file with locations for roads, tiles and its vertices.
Probably needs buttons for placing roads and settlements (and upgrading them).
The nodes have an fx:id that can be used in the controller to refer to them and change them. I've edited and added the tile images. Since the nodes are hexagons, they don't have to be hexagon images.